[java] Rename unusedcode/UnusedModifier to unnecessary/UnnecessaryModifier
