+++ pmd-java/src/main/resources/rulesets/java/unnecessary.xml
@@ -342,4 +342,41 @@ public class Foo {
     </example>
   </rule>
 
+<rule name="UnnecessaryModifier"
+          language="java"
+          since="1.02"
+          message="Avoid modifiers which are implied by the context"
+          class="net.sourceforge.pmd.lang.java.rule.unnecessary.UnnecessaryModifierRule"
+          externalInfoUrl="${pmd.website.baseurl}/rules/java/unnecessary.html#UnnecessaryModifier">
+     <description>
+Fields in interfaces and annotations are automatically `public static final`, and methods are `public abstract`.
+Classes, interfaces or annotations nested in an interface or annotation are automatically `public static`
+(all nested interfaces and annotations are automatically static).
+Nested enums are automatically `static`.
+For historical reasons, modifiers which are implied by the context are accepted by the compiler, but are superfluous.
+     </description>
+        <priority>3</priority>
+     <example>
+ <![CDATA[
+ public @interface Annotation {
+  public abstract void bar(); 		// both abstract and public are ignored by the compiler
+  public static final int X = 0; 	// public, static, and final all ignored
+  public static class Bar {} 		// public, static ignored
+  public static interface Baz {} 	// ditto
+}
+public interface Foo {
+  public abstract void bar(); 		// both abstract and public are ignored by the compiler
+  public static final int X = 0; 	// public, static, and final all ignored
+  public static class Bar {} 		// public, static ignored
+  public static interface Baz {} 	// ditto
+}
+public class Bar {
+  public static interface Baz {} // static ignored
+  public static enum FoorBar { // static ignored
+    FOO;
+  }
+}
+ ]]>
+     </example>
+     </rule>
 </ruleset>
